requirements.txt
streamlit
azure-storage-blob
# pymssql
dotenv
pyodbc